# openapi.yaml
openapi: '3.0.3'
info:
  title: Fotografija API
  description: REST API for managing photoshoots, photos, and comments
  version: '1.0.0'
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://129.151.213.227:3000/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from NextAuth
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
...
paths:
...
  /users:
    get:
      summary: List all users
      description: Returns all users. Admin only.
      operationId: listUsers
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Authentication required
        '403':
          description: Forbidden - Admin role required
...